requirements.txt
streamlit
pandas
scikit-learn
plotly
orion-ml
adtk
altair
plotly
numpy
keras
io